Rename file useRecommendedPlugins to useRecommendedPluginsWithoutChannels.
